Services: a file service is healthy only when the server exports the share, so the probe reads samba's own list and the panel can ask a host for it.

+++ hub/neutrino_hub/utils/subprocess_run.py
@@ -41,7 +41,7 @@ def run(
     command: list[str],
     *,
     input_text: str | None = None,
-    timeout_s: int = DEFAULT_TIMEOUT_S,
+    timeout_s: float = DEFAULT_TIMEOUT_S,
     is_checked: bool = True,
     environment: dict[str, str] | None = None,
 ) -> CommandResult:
